Extract PaymentCard component and refactor history
Add a new PaymentCard composable (app/.../PaymentCard.kt) that encapsulates the payment item UI and formatting helpers (formatColones, formatPaymentMethod). Refactor PaymentHistoryScreen to use PaymentCard instead of the inlined PaymentItem and remove the duplicated UI and formatting functions/imports, cleaning up unused imports and simplifying the screen code for reuse and maintainability.
